SAS-163 reorganizate game files, create GameTable component
